reorganized home component and test into /app
