+++ client/src/components/app/index.ts
@@ -1,4 +1,8 @@
+import About from "./About";
 import Faq from "./Faq";
+import Home from "./home";
 import Members from "./Members";
+import NavBar from "./NavBar";
+import Recruitment from "./Recruitment";
 
-export { Faq, Members };
+export { About, Faq, Home, Members, NavBar, Recruitment };
